# docs/reference/solar_complete_spec_md/07_openapi.yaml
openapi: 3.1.0
info:
  title: Kuozui Green Energy Display API
  version: 1.0.0
  description: API for solar green energy display player using Fastify, MQTT, Socket.IO, and SQLite.
servers:
  - url: /api
paths:
...
  /settings/mqtt:
    get:
      summary: Get MQTT settings
      operationId: getMqttSettings
      responses:
        '200':
          description: MQTT settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MqttSettings'
    put:
      summary: Update MQTT settings
      operationId: updateMqttSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MqttSettingsInput'
      responses:
        '200':
          description: Updated MQTT settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MqttSettings'
...
  /settings/mqtt/topics:
    get:
      summary: Get MQTT topic mappings
      operationId: getTopicMappings
      responses:
        '200':
          description: Topic mappings
          content:
            application/json:
              schema:
                type: object
                properties:
                  topics:
                    type: array
                    items:
                      $ref: '#/components/schemas/TopicMapping'
    put:
      summary: Update MQTT topic mappings
      operationId: updateTopicMappings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                topics:
                  type: array
                  items:
                    $ref: '#/components/schemas/TopicMappingInput'
      responses:
        '200':
          description: Updated topic mappings
...
components:
...
  schemas:
...
    MqttSettings:
      allOf:
        - $ref: '#/components/schemas/MqttSettingsInput'
        - type: object
          properties:
            lastConnectionStatus: { type: string }
            lastMessageAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    MqttSettingsInput:
      type: object
      properties:
        enabled: { type: boolean }
        mode: { type: string, enum: [mqtt, mock] }
        brokerUrl: { type: string }
        port: { type: integer }
        username: { type: string }
        password: { type: string }
        clientId: { type: string }
        reconnectIntervalSec: { type: integer }
        messageTimeoutSec: { type: integer }
    TopicMapping:
      allOf:
        - $ref: '#/components/schemas/TopicMappingInput'
        - type: object
          properties:
            id: { type: string }
            updatedAt: { type: string, format: date-time }
    TopicMappingInput:
      type: object
      properties:
        metricKey: { type: string }
        labelZh: { type: string }
        labelEn: { type: string }
        topic: { type: string }
        unit: { type: string }
        enabled: { type: boolean }
...